backend_api/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
python-jose[cryptography]>=3.3.0
passlib>=1.7.4
pydantic-settings>=2.5.0
python-multipart>=0.0.9
streamlit>=1.28.0
pandas>=2.0.0
openpyxl>=3.1.0
python-dateutil>=2.8.2
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
